feat: support windows signal
